Move FFmpeg code from ffmpeg.c to ffmpeg_decoder.c for consistency
